think finish change pal types to u16 *

+++ include/battle_anim.h
@@ -783,7 +783,7 @@ void AnimTask_GetAttackerSide(u8 taskId);
 void AnimTask_GetTargetSide(u8 taskId);
 void AnimTask_GetTargetIsAttackerPartner(u8 taskId);
 void AnimTask_SetAllNonAttackersInvisiblity(u8 taskId);
-void sub_80BBA20(u8 taskId, s32 unused, u16 arg2, u8 battler1, u8 arg4, u8 arg5, u8 arg6, u8 arg7, const u32 *gfx, const u32 *tilemap, const u32 *palette);
+void sub_80BBA20(u8 taskId, s32 unused, u16 arg2, u8 battler1, u8 arg4, u8 arg5, u8 arg6, u8 arg7, const u32 *gfx, const u32 *tilemap, const u16 *palette);
 void AnimTask_GetBattleTerrain(u8 taskId);
 void AnimTask_AllocBackupPalBuffer(u8 taskId);
 void AnimTask_FreeBackupPalBuffer(u8 taskId);
